Node that creates ring is its own predecessor.
Consequently, I had to make check_predecessor() not ping a node's
predecessor if said predecessor IS that node.
Furthermore, when a node creates a ring, I initialize its finger table
such that each finger entry points to that node; in other words, if a
node is the only node in the ring, I think its fingers should all point
to itself.